requirements.txt
azure-cosmos==3.1.1